split util into two
